[luci] move template function to new head file
move use_paddings funtion to new .h file

ONE-DCO-1.0-Signed-off-by: JuYoung Lee rsb98759@gmail.com
