+++ compiler/luci/service/src/Nodes/CirclePad.cpp
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
+#include "luci/Service/CircleShapeInference.h"
+
 #include "CircleCloneNode.h"
 #include "CircleShapeInferenceHelper.h"
-
-#include "luci/Service/CircleShapeInference.h"
+#include "CircleShapeInferenceHelperPads.h"
 
 namespace luci
 {
